+++ app/Http/Controllers/ReservationController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Auth;
 
 use Carbon\Carbon;
 
@@ -17,6 +18,8 @@ class ReservationController extends Controller
      */
     public function index(ReservableItem $item)
     {
+        $this->authorize('viewAny', ReservableItem::class);
+
         return response()->json(
             Reservation::where('reservable_item_id', $item->id)
                          ->orderBy('reserved_from')
@@ -29,6 +32,8 @@ class ReservationController extends Controller
      */
     public function indexForWashingMachines()
     {
+        $this->authorize('viewAny', ReservableItem::class);
+
         $items = ReservableItem::where('type', 'washing_machine')->get();
         $from = Carbon::today()->startOfWeek();
         $until = $from->copy()->addDays(7);
@@ -47,6 +52,8 @@ class ReservationController extends Controller
      */
     public function show(Reservation $reservation)
     {
+        $this->authorize('view', $reservation);
+
         return view('reservations.show', [
             'reservation' => $reservation
         ]);
@@ -57,7 +64,28 @@ class ReservationController extends Controller
      */
     public function create(ReservableItem $item)
     {
-        abort(500, 'create not implemented yet');
+        $this->authorize('requestReservation', $item);
+
+        return view('reservations.edit', [
+            'item' => $item
+        ]);
+    }
+
+    /**
+     * Aborts the request if there is already a reservation
+     * which would conflict with the one given.
+     * Note: we assume that this reservation is not yet saved
+     */
+    public static function abortConflictingReservation(Reservation $newReservation)
+    {
+        $conflictingReservations = $newReservation->reservableItem
+            ->reservationsInSlot($newReservation->reserved_from, $newReservation->reserved_until);
+
+        if (!$conflictingReservations->empty()) {
+            abort(409, "Reservation already exists in the given interval:
+                {$conflictingReservations->first()->reserved_from},
+                {$conflictingReservations->first()->reserved_until}");
+        }
     }
 
     /**
@@ -67,6 +95,8 @@ class ReservationController extends Controller
      */
     public function store(ReservableItem $item, Request $request)
     {
+        $this->authorize('requestReservation', $item);
+
         $validator = Validator::make($request->all(), [
             'title' => 'nullable|string|max:255',
             'note' => 'nullable|string|max:2047',
@@ -82,24 +112,38 @@ class ReservationController extends Controller
         $newReservation->user_id = user()->id;
         $newReservation->title = $validatedData['title'];
         $newReservation->note = $validatedData['note'];
-        $newReservation->reserved_from = $validatedData['reserved_from'];
-        $newReservation->reserved_until = $validatedData['reserved_until'];
+        $newReservation->reserved_from = Carbon::make($validatedData['reserved_from']);
+        $newReservation->reserved_until = Carbon::make($validatedData['reserved_until']);
 
-        if (!empty($item->reservationsInSlot($newReservation->reserved_from, $newReservation->reserved_until))) {
-            abort(409, 'Reservation already exists in the given interval');
-        }
+        $newReservation->verified = Auth::user()->can('reserveImmediately', $item);
+
+        ReservationController::abortConflictingReservation($newReservation);
 
         // and finally:
         $newReservation->save();
 
-        return response()->json($newReservation);
+        return route('reservations.show', $newReservation);
+    }
+
+    /**
+     * Returns a form for creating a reservation.
+     */
+    public function edit(Reservation $reservation)
+    {
+        $this->authorize('modify', $reservation);
+
+        return view('reservations.edit', [
+            'reservation' => $reservation
+        ]);
     }
 
     /**
      * Deletes a reservation.
      */
     public function delete(Reservation $reservation) {
+        $this->authorize('modify', $reservation);
+
         $reservation->delete();
-        return redirect()->back()->with('message', __('general.successful_modification'));
+        return redirect()->route('reservations.items.show', $reservation->reservableItem);
     }
 }
